Fix datatype for ids in every entitny

+++ migrations/004_create_projects_table.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE projects (
-    id BIGINT AUTO_INCREMENT PRIMARY KEY,
-    client_id BIGINT NOT NULL,
+    id INT AUTO_INCREMENT PRIMARY KEY,
+    client_id INT NOT NULL,
     country VARCHAR(100) NOT NULL,
     services_needed JSON NOT NULL,
     budget DECIMAL(12,2),
@@ -8,4 +8,4 @@ CREATE TABLE projects (
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
     FOREIGN KEY (client_id) REFERENCES clients(id)
-);
+);
